+++ apps/pixi-js-demo/emp.config.ts
@@ -1,0 +1,10 @@
+import {defineConfig} from '@empjs/cli'
+
+export default defineConfig(() => ({
+  chain(config) {
+    config.performance.hints('warning').maxAssetSize(512 * 1024)
+  },
+  server: {
+    open: false,
+  },
+}))