+++ projects/esm-react-app/@mf-types/esmReactHost/App.d.ts
@@ -1,2 +1,0 @@
-export * from './compiled-types/App';
-export { default } from './compiled-types/App';