+++ projects/esm-react-app/@mf-types/esmReactHost/apis.d.ts
@@ -1,3 +1,0 @@
-
-    export type RemoteKeys = 'esmReactHost/App';
-    type PackageType<T> = T extends 'esmReactHost/App' ? typeof import('esmReactHost/App') :any;