+++ projects/esm-react-app/@mf-types/esmReactHost/compiled-types/App.d.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import './App.css';
-declare function App(): React.JSX.Element;
-export default App;